+++ src/main/java/com/example/primo_progetto_spring/Entity/Exercise.java
@@ -1,6 +1,6 @@
 package com.example.primo_progetto_spring.Entity;
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.example.primo_progetto_spring.Student.Studente;
 import jakarta.persistence.*;
 
 //Aggiungere una nuova entità "Exercise", definita da id, id dello studente, consegna (stringa) e voto (intero).
