+++ src/main/java/com/example/primo_progetto_spring/Service/TutorService.java
@@ -4,7 +4,7 @@ import com.example.primo_progetto_spring.Entity.Classroom;
 import com.example.primo_progetto_spring.Entity.Tutor;
 import com.example.primo_progetto_spring.dto.AddTutorToClassroom;
 import com.example.primo_progetto_spring.repository.ClassroomRepository;
-import com.example.primo_progetto_spring.repository.StudenteRepository;
+import com.example.primo_progetto_spring.Student.StudenteRepository;
 import com.example.primo_progetto_spring.repository.TutorRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
