+++ src/main/java/com/example/primo_progetto_spring/Student/Studente.java
@@ -1,6 +1,8 @@
-package com.example.primo_progetto_spring.Entity;
+package com.example.primo_progetto_spring.Student;
 
 
+import com.example.primo_progetto_spring.Entity.Classroom;
+import com.example.primo_progetto_spring.Entity.Exercise;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import jakarta.persistence.*;
