+++ src/main/java/com/example/primo_progetto_spring/Student/StudenteRepository.java
@@ -1,6 +1,5 @@
-package com.example.primo_progetto_spring.repository;
+package com.example.primo_progetto_spring.Student;
 
-import com.example.primo_progetto_spring.Entity.Studente;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -12,7 +11,7 @@ public interface StudenteRepository extends JpaRepository<Studente, Long>{
 
     List<Studente> findByNome (String nome);
 
-    List<Studente> findByNomeStartingWith (String prefisso);
+    List<Studente> findByNomeStartingWith(String prefisso);
 
     List<Studente> findByDataLessThanEqual(LocalDate dataMinima);
 }
